+++ lint-staged.config.js
@@ -1,3 +1,3 @@
 import config from '@airbnb-extended/lint-staged-config/root';
 
-export default { ...config };
+export default config;