+++ packages/eslint-config-airbnb-extended/eslint.config.js
@@ -1,4 +1,4 @@
-import config from '../../configs/eslint-config/src/eslint-internal.config.js';
-import { defineConfig } from '../../configs/eslint-config/src/eslint-utils.js';
+import config from '@airbnb-extended/eslint-config/base';
+import { defineConfig } from '@airbnb-extended/eslint-config/utils';
 
 export default defineConfig(config);